chore: skip plugin loading if tar is empty

+++ src/managers/connect.py
@@ -128,6 +128,13 @@ class ConnectManager:
             logger.debug(f"Plugin {resource_path.name} already loaded, skipping...")
             return
 
+        # the checksum here corresponds to a sha256sum of an empty tar files created using --files-from /dev/null
+        # this is what is loaded by default as the connect-plugin resource in Charmhub
+        # as Charmhub won't allow charms without a resource
+        if self._plugin_checksum(resource_path) == "84ff92691f909a05b224e1c56abb4864f01b4f8e3c854e4bb4c7baf1d3f6d652":
+            logger.debug("Plugin is empty, skipping...")
+            return
+
         load_path = self._create_plugin_dir(resource_path, path_prefix=path_prefix)
         self._untar_plugin(resource_path, load_path)
         self.workload.rmdir(f"{resource_path}")
